Add Release(ResourceHandle handle) API in ResourceSystem

+++ Modules/Resource/ResourceHandle.cs
@@ -12,8 +12,8 @@ namespace Kurisu.Framework.Resource
     /// </summary>
     public readonly struct ResourceHandle : IEquatable<ResourceHandle>
     {
-        internal readonly int handleID;
-        internal readonly int operationType;
+        internal readonly uint handleID;
+        internal readonly byte operationType;
         internal readonly AsyncOperationHandle InternalHandle => ResourceSystem.CastOperationHandle(handleID);
         public readonly object Result => InternalHandle.Result;
 #if UNITASK_SUPPORT
@@ -21,7 +21,7 @@ namespace Kurisu.Framework.Resource
 #else
         public readonly Task Task => InternalHandle.Task;
 #endif
-        public ResourceHandle(int handleID, int operationType)
+        public ResourceHandle(uint handleID, byte operationType)
         {
             this.handleID = handleID;
             this.operationType = operationType;
@@ -48,8 +48,8 @@ namespace Kurisu.Framework.Resource
     /// </summary>
     public readonly struct ResourceHandle<T> : IEquatable<ResourceHandle<T>>
     {
-        internal readonly int handleID;
-        internal readonly int operationType;
+        internal readonly uint handleID;
+        internal readonly byte operationType;
         internal readonly AsyncOperationHandle<T> InternalHandle => ResourceSystem.CastOperationHandle<T>(handleID);
         public readonly T Result => InternalHandle.Result;
 #if UNITASK_SUPPORT
@@ -57,7 +57,7 @@ namespace Kurisu.Framework.Resource
 #else
         public readonly Task<T> Task => InternalHandle.Task;
 #endif
-        public ResourceHandle(int handleID, int operationType)
+        public ResourceHandle(uint handleID, byte operationType)
         {
             this.handleID = handleID;
             this.operationType = operationType;
